(refactor) Reorganize things to match naming conventions
