+++ .eslintrc
@@ -1,4 +1,32 @@
 {
+  "extends": ["ts-react-important-stuff", "plugin:prettier/recommended"],
   "parser": "@typescript-eslint/parser",
-  "extends": ["ts-react-important-stuff", "plugin:prettier/recommended"]
+  "ignorePatterns": ["**/*.test.tsx"],
+  "rules": {
+    "no-restricted-imports": [
+      "error",
+      {
+        "paths": [
+          {
+            "name": "lodash",
+            "message": "Import specific methods from `lodash-es`. e.g. `import map from 'lodash-es/map'`"
+          },
+          {
+            "name": "lodash-es",
+            "message": "Import specific methods from `lodash-es`. e.g. `import map from 'lodash-es/map'`"
+          }
+        ],
+        "patterns": [
+          {
+            "group": ["carbon-components-react"],
+            "message": "Import from `@carbon/react` directly. e.g. `import { Toggle } from '@carbon/react'`"
+          },
+          {
+            "group": ["@carbon/icons-react"],
+            "message": "Import from `@carbon/react/icons`. e.g. `import { ChevronUp } from '@carbon/react/icons'`"
+          }
+        ]
+      }
+    ]
+  }
 }
