+++ src/components/Panel/panel.component.tsx
