+++ src/components/reporting-item/reporting-item.component.tsx
